fix: fix error of mismatch of types passed to onChange
